+++ paramak/__init__.py
@@ -2,6 +2,7 @@ from .shape import Shape
 from .reactor import Reactor
 from .utils import rotate, extend, distance_between_two_points, diff_between_angles
 from .utils import EdgeLengthSelector, FaceAreaSelector
+from .neutronics_utils import define_moab_core_and_tags, add_stl_to_moab_core
 
 from .parametric_shapes.extruded_mixed_shape import ExtrudeMixedShape
 from .parametric_shapes.extruded_spline_shape import ExtrudeSplineShape
